MODSOURCE-375 Remove unnecessary handler for marc holdings
- removed MarcHoldingsEventHandler and AbstractMarcEventHandler
- reverted MarcAuthorityEventHandler to previous commit
- refactored AbstractMarcEventHandlerTest to MarcEventHandlerTest
