+++ mod-source-record-storage-server/src/test/java/org/folio/services/MarcEventHandlerTest.java
@@ -2,7 +2,6 @@ package org.folio.services;
 
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.MATCH_PROFILE;
 import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_AUTHORITY;
-import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_HOLDING;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -36,25 +35,21 @@ import org.folio.rest.jaxrs.model.ProfileSnapshotWrapper;
 import org.folio.rest.jaxrs.model.RawRecord;
 import org.folio.rest.jaxrs.model.Record;
 import org.folio.rest.jaxrs.model.Snapshot;
-import org.folio.services.handlers.AbstractMarcEventHandler;
 import org.folio.services.handlers.MarcAuthorityEventHandler;
-import org.folio.services.handlers.MarcHoldingsEventHandler;
 
 @RunWith(VertxUnitRunner.class)
-public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
+public class MarcEventHandlerTest extends AbstractLBServiceTest {
 
   private static final String NOT_VALID_ENTITY_TYPE = "notValidEntityType";
   private static final String PARSED_CONTENT_WITH_ADDITIONAL_FIELDS = "{\"leader\":\"01589ccm a2200373   4500\",\"fields\":[{\"245\":{\"ind1\":\"1\",\"ind2\":\"0\",\"subfields\":[{\"a\":\"Neue Ausgabe sämtlicher Werke,\"}]}},{\"948\":{\"ind1\":\"\",\"ind2\":\"\",\"subfields\":[{\"a\":\"acf4f6e2-115c-4509-9d4c-536c758ef917\"},{\"b\":\"681394b4-10d8-4cb1-a618-0f9bd6152119\"},{\"d\":\"12345\"},{\"e\":\"lts\"},{\"x\":\"addfast\"}]}},{\"999\":{\"ind1\":\"f\",\"ind2\":\"f\",\"subfields\":[{\"s\":\"bc37566c-0053-4e8b-bd39-15935ca36894\"}]}}]}";
-  private static final String PAYLOAD_HAS_NO_DATA_MSG = "Failed to handle event payload, cause event payload context does not contain MARC data";
+  private static final String PAYLOAD_HAS_NO_DATA_MSG = "Failed to handle event payload, cause event payload context does not contain MARC_AUTHORITY data";
   private static final String RECORD_ID = "acf4f6e2-115c-4509-9d4c-536c758ef917";
   private static final String SNAPSHOT_ID_1 = UUID.randomUUID().toString();
   private static final String SNAPSHOT_ID_2 = UUID.randomUUID().toString();
   private static RawRecord rawRecord;
   private RecordDao recordDao;
   private Record marcAuthorityRecord;
-  private Record marcHoldingsRecord;
   private MarcAuthorityEventHandler marcAuthorityEventHandler;
-  private MarcHoldingsEventHandler marcHoldingsEventHandler;
 
   @BeforeClass
   public static void setUpClass() throws IOException {
@@ -68,7 +63,6 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
 
     recordDao = new RecordDaoImpl(postgresClientFactory);
     marcAuthorityEventHandler = new MarcAuthorityEventHandler();
-    marcHoldingsEventHandler = new MarcHoldingsEventHandler();
     Async async = context.async();
 
     Snapshot snapshot1 = new Snapshot()
@@ -99,21 +93,6 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
         .withInstanceHrid("12345"))
       .withState(Record.State.ACTUAL);
 
-    this.marcHoldingsRecord = new Record()
-      .withId(RECORD_ID)
-      .withMatchedId(RECORD_ID)
-      .withSnapshotId(SNAPSHOT_ID_1)
-      .withGeneration(0)
-      .withRecordType(MARC_HOLDING)
-      .withRawRecord(rawRecord)
-      .withParsedRecord(new ParsedRecord()
-        .withId(RECORD_ID)
-        .withContent(PARSED_CONTENT_WITH_ADDITIONAL_FIELDS))
-      .withExternalIdsHolder(new ExternalIdsHolder()
-        .withInstanceId("681394b4-10d8-4cb1-a618-0f9bd6152119")
-        .withInstanceHrid("12345"))
-      .withState(Record.State.OLD);
-
     SnapshotDaoUtil.save(postgresClientFactory.getQueryExecutor(TENANT_ID), snapshots).onComplete(save -> {
       if (save.failed()) {
         context.fail(save.cause());
@@ -136,12 +115,32 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
 
   @Test
   public void shouldHandleMarcAuthorityRecord(TestContext context) {
-    shouldHandleMarcRecord(context, marcAuthorityEventHandler, EntityType.MARC_AUTHORITY, marcAuthorityRecord);
-  }
+    Async async = context.async();
 
-  @Test
-  public void shouldHandleMarcHoldingsRecord(TestContext context) {
-    shouldHandleMarcRecord(context, marcHoldingsEventHandler, EntityType.MARC_HOLDINGS, marcHoldingsRecord);
+    HashMap<String, String> payloadContext = new HashMap<>();
+    payloadContext.put(EntityType.MARC_AUTHORITY.value(), Json.encode(marcAuthorityRecord));
+
+    DataImportEventPayload dataImportEventPayload = new DataImportEventPayload()
+      .withContext(payloadContext)
+      .withTenant(TENANT_ID)
+      .withCurrentNode(new ProfileSnapshotWrapper()
+        .withId(UUID.randomUUID().toString())
+        .withContentType(MATCH_PROFILE));
+
+    recordDao.saveRecord(marcAuthorityRecord, TENANT_ID)
+      .onComplete(context.asyncAssertSuccess())
+      .onSuccess(record -> marcAuthorityEventHandler.handle(dataImportEventPayload)
+        .whenComplete((updatedEventPayload, throwable) -> {
+          context.assertNull(throwable);
+          context.assertEquals(1, updatedEventPayload.getEventsChain().size());
+          var actualRecord = new JsonObject(updatedEventPayload.getContext().get(EntityType.MARC_AUTHORITY.value())).mapTo(Record.class);
+          context.assertEquals(actualRecord.getId(), record.getId());
+          context.assertEquals(actualRecord.getSnapshotId(), record.getSnapshotId());
+          context.assertEquals(actualRecord.getRawRecord(), record.getRawRecord());
+          context.assertEquals(actualRecord.getRecordType(), record.getRecordType());
+          context.assertEquals(actualRecord.getErrorRecord(), record.getErrorRecord());
+          async.complete();
+        }));
   }
 
   @Test
@@ -150,18 +149,8 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
   }
 
   @Test
-  public void shouldNotHandleMarcHoldingsRecordIfContextIsNull(TestContext context) {
-    shouldNotHandleMarcRecord(context, marcHoldingsEventHandler, marcHoldingsRecord, null);
-  }
-
-  @Test
   public void shouldNotHandleMarcAuthorityRecordIfContextIsEmpty(TestContext context) {
     shouldNotHandleMarcRecord(context, marcAuthorityEventHandler, marcAuthorityRecord, new HashMap<>());
-  }
-
-  @Test
-  public void shouldNotHandleMarcHoldingsRecordIfContextIsEmpty(TestContext context) {
-    shouldNotHandleMarcRecord(context, marcHoldingsEventHandler, marcHoldingsRecord, new HashMap<>());
   }
 
   @Test
@@ -172,20 +161,8 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
   }
 
   @Test
-  public void shouldNotHandleMarcHoldingsRecordIfContextIsNotEqualEntityType(TestContext context) {
-    HashMap<String, String> payloadContext = new HashMap<>();
-    payloadContext.put(NOT_VALID_ENTITY_TYPE, Json.encode(marcHoldingsRecord));
-    shouldNotHandleMarcRecord(context, marcHoldingsEventHandler, marcHoldingsRecord, payloadContext);
-  }
-
-  @Test
   public void shouldReturnTrueIfMarcAuthorityPayloadIsEligible(TestContext context) {
     shouldReturnMarcPayloadIsEligible(context, EntityType.MARC_AUTHORITY.value(), marcAuthorityEventHandler, marcAuthorityRecord, true);
-  }
-
-  @Test
-  public void shouldReturnTrueIfMarcHoldingsPayloadIsEligible(TestContext context) {
-    shouldReturnMarcPayloadIsEligible(context, EntityType.MARC_HOLDINGS.value(), marcHoldingsEventHandler, marcHoldingsRecord, true);
   }
 
   @Test
@@ -194,17 +171,11 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
   }
 
   @Test
-  public void shouldReturnFalseIfMarcHoldingsPayloadIsEligible(TestContext context) {
-    shouldReturnMarcPayloadIsEligible(context, NOT_VALID_ENTITY_TYPE, marcHoldingsEventHandler, marcHoldingsRecord, false);
-  }
-
-  @Test
   public void shouldReturnFalseIsPostProcessingNeeded(TestContext context) {
     context.assertFalse(marcAuthorityEventHandler.isPostProcessingNeeded());
-    context.assertFalse(marcHoldingsEventHandler.isPostProcessingNeeded());
   }
 
-  private void shouldReturnMarcPayloadIsEligible(TestContext context, String entityType, AbstractMarcEventHandler marcEventHandler, Record marcRecord, boolean flag) {
+  private void shouldReturnMarcPayloadIsEligible(TestContext context, String entityType, MarcAuthorityEventHandler marcEventHandler, Record marcRecord, boolean flag) {
     HashMap<String, String> payloadContext = new HashMap<>();
     payloadContext.put(entityType, Json.encode(marcRecord));
 
@@ -214,37 +185,7 @@ public class AbstractMarcEventHandlerTest extends AbstractLBServiceTest {
     context.assertEquals(marcEventHandler.isEligible(dataImportEventPayload), flag);
   }
 
-  private void shouldHandleMarcRecord(TestContext context, AbstractMarcEventHandler marcEventHandler,
-                                      EntityType entityType, Record marcRecord) {
-    Async async = context.async();
-
-    HashMap<String, String> payloadContext = new HashMap<>();
-    payloadContext.put(entityType.value(), Json.encode(marcRecord));
-
-    DataImportEventPayload dataImportEventPayload = new DataImportEventPayload()
-      .withContext(payloadContext)
-      .withTenant(TENANT_ID)
-      .withCurrentNode(new ProfileSnapshotWrapper()
-        .withId(UUID.randomUUID().toString())
-        .withContentType(MATCH_PROFILE));
-
-    recordDao.saveRecord(marcRecord, TENANT_ID)
-      .onComplete(context.asyncAssertSuccess())
-      .onSuccess(record -> marcEventHandler.handle(dataImportEventPayload)
-        .whenComplete((updatedEventPayload, throwable) -> {
-          context.assertNull(throwable);
-          context.assertEquals(1, updatedEventPayload.getEventsChain().size());
-          var actualRecord = new JsonObject(updatedEventPayload.getContext().get(entityType.value())).mapTo(Record.class);
-          context.assertEquals(actualRecord.getId(), record.getId());
-          context.assertEquals(actualRecord.getSnapshotId(), record.getSnapshotId());
-          context.assertEquals(actualRecord.getRawRecord(), record.getRawRecord());
-          context.assertEquals(actualRecord.getRecordType(), record.getRecordType());
-          context.assertEquals(actualRecord.getErrorRecord(), record.getErrorRecord());
-          async.complete();
-        }));
-  }
-
-  private void shouldNotHandleMarcRecord(TestContext context, AbstractMarcEventHandler marcEventHandler,
+  private void shouldNotHandleMarcRecord(TestContext context, MarcAuthorityEventHandler marcEventHandler,
                                          Record marcRecord, HashMap<String, String> payloadContext) {
     Async async = context.async();
